ProductWeb_RestAPI add function findProductById, add, delete product

+++ ProductWeb_RestAPI/src/main/java/com/thinne/backend/business/Bean/ProductBean.java
@@ -25,7 +25,7 @@ public class ProductBean implements ProductLocal {
     }
 
     @Override
-    public Product getProduct(long id) {
+    public Product getProduct(int id) {
         return productRepository.findById(id);
     }
